v1.152:
(included all changes since v1.151)

HIGHLIGHTS:
- improvements and bug fixes

BREAKING CHANGES:
- Core: changed "BarrierGroupDesc" layout to {data, num, data, num, data, num} (to match overall style)
- Core: "usageMask" renamed to "usage"
- Core: "typedBufferOffsetAlignment" and "storageBufferOffsetAlignment" merged into a single "bufferShaderResourceOffsetAlignment" in "DeviceDesc" (in any case they should match)
- RayTracing: "RAY_TRACING_BUFFER" decoupled into "SCRATCH_BUFFER" and "SHADER_BINDING_TABLE" (was an umbrella, hiding potential inefficiencies)
- RayTracing: "ACCELERATION_STRUCTURE_BUILD_READ" renamed to "ACCELERATION_STRUCTURE_BUILD_INPUT"
- RayTracing: "rayTracingShaderTableAlignment" renamed to "shaderBindingTableAlignment" (to match overall style)
- RayTracing: "rayTracingScratchAlignment" renamed to "scratchBufferOffsetAlignment" (to match overall style)

DETAILS:
- Core: Added missing "NONE" to "nriGetGraphicsAPIString"
- Core: "ReportMessage" improvement - "messageType, "file" and "line" are not applied to a message text, they can be used for formatting inside a user provided message callback
- Core: "ReportMessage" improvement - the default message callback is modified to reflect the formatting changes
- Core: allow more 0 in "TextureDesc" (will be silently promoted to 1)
- Core: removed ~useless "fragmentShaderCombinedOutputResourceMaxNum" from "DeviceDesc"
- Core: added comments, clarifying compatible "AccessBits" expected by "Layout"
- VK: add missing VK location for input attributes
- ResourceAllocator: auto respected buffer placement alignment VK requirements according to the provided buffer "usage"
- D3D12: properly (?) used "PlaneSlice"
- None: properly initialized "DeviceDesc"
- Helper: improved "FitPipelineLayoutSettingsIntoDeviceLimits"
- RayTracing: increased robustness
- Validation: improved validation of "CreatePipelineLayout" by utilizing "FitPipelineLayoutSettingsIntoDeviceLimits" helper
- Validation: fixed and improved incomplete validation for "AllocateDescriptorSets"
- code polishing

+++ Include/Extensions/NRIRayTracing.h
@@ -21,7 +21,7 @@ NriEnum(CopyMode, uint8_t,
     COMPACT
 );
 
-NriBits(BottomLevelGeometryBits, uint32_t,
+NriBits(BottomLevelGeometryBits, uint8_t,
     NONE                                = 0,
     OPAQUE_GEOMETRY                     = NriBit(0),
     NO_DUPLICATE_ANY_HIT_INVOCATION     = NriBit(1)
@@ -35,7 +35,7 @@ NriBits(TopLevelInstanceBits, uint32_t,
     FORCE_NON_OPAQUE                    = NriBit(3)
 );
 
-NriBits(AccelerationStructureBuildBits, uint32_t,
+NriBits(AccelerationStructureBuildBits, uint8_t,
     NONE                                = 0,
     ALLOW_UPDATE                        = NriBit(0),
     ALLOW_COMPACTION                    = NriBit(1),
